expgfx_addremove: int slotType param (was short) drops redundant extsh on ==1 compare, matches retail r24 (90.85->91.02); register-neutral header+def, no direct callers

+++ include/main/dll_000A_expgfx.h
@@ -30,7 +30,7 @@ void expgfx_free(u32 sourceId);
 void expgfx_resetAllPools(void);
 void expgfx_updateFrameState(int sourceMode,int sourceId);
 struct ExpgfxSpawnConfig;
-int expgfx_addremove(struct ExpgfxSpawnConfig *config, int preferredPoolIndex, short slotType,
+int expgfx_addremove(struct ExpgfxSpawnConfig *config, int preferredPoolIndex, int slotType,
                      u8 boundsTemplateId);
 void expgfx_onMapSetup(void);
 void expgfx_release(void);
